requirements.txt
psycopg2-binary==2.9.9
pandas==2.0.3
openpyxl==3.1.2
reportlab==4.0.4
Pillow==10.0.0